feat: organize the mock data generation a bit better
