+++ apps/react-ui/client/src/utils/dataUtils.ts
@@ -1,33 +1,8 @@
 import * as XLSX from "xlsx"
-import { faker } from "@faker-js/faker"
 
 // Generate a unique ID for uploaded data
 export const generateDataId = (): string => {
 	return `data_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`
-}
-
-// Generate mock CSV data for development
-export const generateMockCSVFile = (): File => {
-	// Generate 10-20 rows of realistic data
-	const numRows = faker.number.int({ min: 10, max: 20 })
-
-	// Create CSV content
-	const headers = ["effect", "se", "n_obs", "study_id"]
-	const csvRows = [headers.join(",")]
-
-	for (let i = 0; i < numRows; i++) {
-		const effect = faker.number.float({ min: -2, max: 2, multipleOf: 0.001 })
-		const se = faker.number.float({ min: 0.01, max: 0.5, multipleOf: 0.001 })
-		const nObs = faker.number.int({ min: 50, max: 1000 })
-		const studyId = i + 1
-
-		csvRows.push(`${effect},${se},${nObs},${studyId}`)
-	}
-
-	const csvContent = csvRows.join("\n")
-	const blob = new Blob([csvContent], { type: "text/csv" })
-
-	return new File([blob], "mock_data.csv", { type: "text/csv" })
 }
 
 // Process uploaded file and return structured data
